requirements.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
heroku==0.1.4
idna==3.4
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.7
pypng==0.20220715.0
python-dateutil==1.5
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
